requirements.txt
boto3==1.10.10
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
-e git://github.com/arneb/django-messages.git#egg=django_messages
django-storages==1.7.2
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
django-crispy-forms==1.8.0
